--- src/app/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "invariant-app"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-gstd.workspace = true
-traceable_result.workspace = true
-math.workspace = true
-decimal.workspace = true
-io.workspace = true
-contracts.workspace = true
-gear-core.workspace = true
-sails-rtl.workspace = true
-futures = { version = "0.3.30", features = [
-  "async-await",
-], default-features = false }
-
-[dev-dependencies]
-env_logger.workspace = true
-gclient.workspace = true
-gtest.workspace = true
-gear-core.workspace = true
-tokio = "1"
-
-[build-dependencies]
-gear-wasm-builder.workspace = true
-
-[features]
-default = []
-test = ["gstd/debug"]
