+++ src/invariant_storage.rs
