+++ src/test_helpers/gtest/entrypoints/deposit.rs
@@ -3,7 +3,7 @@ use gtest::*;
 use io::*;
 use math::{sqrt_price::SqrtPrice, token_amount::TokenAmount};
 use sails_rtl::ActorId;
-
+use gstd::{vec, ToString, String};
 use crate::{send_request, test_helpers::gtest::InvariantResult};
 
 #[track_caller]
