+++ src/test_helpers/gtest/mod.rs
@@ -11,3 +11,4 @@ pub use entrypoints::*;
 pub use snippets::*;
 pub use token::*;
 pub use utils::*;
+pub use gstd::{Vec, vec, ToOwned};
