+++ src/wasm/src/lib.rs
@@ -2,4 +2,4 @@
 
 #[allow(unused_imports)]
 #[cfg(target_arch = "wasm32")]
-pub use invariant_app::wasm::*;
+pub use invariant::wasm::*;
